Update views.py
change in line 45 48 (unnecessary else after a return.

+++ circle/views.py
@@ -41,11 +41,11 @@ def addPerson(request):
 				print(person)
 				Chat.objects.create(left=person, right=person)
 				return HttpResponseRedirect(reverse('person', args=(person.id, )))
-			else:
-				return render(request, "circle/error.html", context={"message": "Invalid Data.!!", "type": "Type Error", "link": "search"})
-		else:
-			form = PersonForm(instance=person)
-			return render(request, "circle/person.html", context={"person": person, "form": form})
+			
+			return render(request, "circle/error.html", context={"message": "Invalid Data.!!", "type": "Type Error", "link": "search"})
+		
+		form = PersonForm(instance=person)
+		return render(request, "circle/person.html", context={"person": person, "form": form})
 	except Exception as e:
 		return render(request, "circle/error.html", context={"message": str(e), "type": "Internal Error", "link": "newPerson"})
 
@@ -148,16 +148,13 @@ def edit(request, id, type):
 				if form.is_valid():
 					form.save()
 					return HttpResponseRedirect(reverse('tag', args=(id, )))
-				else:
-					return render(request, "circle/error.html", context={"message": "Invalid Data.!!", "type": "Type Error", "link": "tags"})
-			else:
-				form = TagForm(instance=tag)
-				return render(request, "circle/tag.html", context={"tag": tag, "form": form})
+				return render(request, "circle/error.html", context={"message": "Invalid Data.!!", "type": "Type Error", "link": "tags"})
+			form = TagForm(instance=tag)
+			return render(request, "circle/tag.html", context={"tag": tag, "form": form})
 		except Tag.DoesNotExist:
 			return render(request, "circle/error.html", context={"message": "No Tag Found.!!", "type": "Type Error", "link": "tags"})
 
-	else:
-		return render(request, "circle/error.html", context={"message": "Invalid Type.!!", "type": "Type Error", "link": "search"})
+	return render(request, "circle/error.html", context={"message": "Invalid Type.!!", "type": "Type Error", "link": "search"})
 
 
 @login_required
@@ -259,9 +256,8 @@ def result(request, type):
 	try:
 		if request.GET:
 			return HttpResponseRedirect(reverse('search', args=(type,)))
-		else:
-			try:
-				search = str(request.POST.get("search"))
+		try:
+			search = str(request.POST.get("search"))
 			except KeyError:
 				return render(request, "circle/error.html", context={"message":  "Enter text to search.!!", "type": "Key Error", "link": "search"})
 			except ValueError:
@@ -283,7 +279,7 @@ def result(request, type):
 				# .exclude(display)
 				# .exclude(rent).exclude(purchased).exclude(sold)
 				return render(request, "circle/result.html", context={'articles': articles, 'type': type})
-			elif type == 'person':
+			if type == 'person':
 				friends = list(Person.objects.get(pk=request.user.person.id).friends.all())
 				persons = Person.objects.filter(Q(first__contains=search) | Q(last__contains=search) | Q(username__contains=search)).exclude(friends__in=friends)
 				return render(request, "circle/result.html", context={'persons': persons, 'type': type})
@@ -329,8 +325,8 @@ def rent(request, article_id):
 			person.rented.add(article)
 			person.save()
 			return HttpResponseRedirect(reverse('rented', args=()))
-		else:
-			return HttpResponseRedirect(reverse('article', args=(article_id, )))
+		
+		return HttpResponseRedirect(reverse('article', args=(article_id, )))
 	except Article.DoesNotExist:
 		return render(request, "circle/error.html", context={"message": "No Article Found.!!", "type": "Type Error", "link": "search"})
 	except Person.DoesNotExist:
@@ -497,25 +493,25 @@ def remove(request, id, type):
 				person.bookmarked.remove(article)
 				person.save()
 				return HttpResponseRedirect(reverse('wishlisted', args=()))
-			elif type == 'rented':
+			if type == 'rented':
 				person.rented.remove(article)
 				person.save()
 				return HttpResponseRedirect(reverse('rented', args=()))
-			elif type == 'carted':
+			if type == 'carted':
 				person.carted.remove(article)
 				person.save()
 				return HttpResponseRedirect(reverse('carted', args=()))
-			# elif type == 'purchased':
-			# 	person.purchased.remove(article)
-			# 	person.save()
-			# 	return HttpResponseRedirect(reverse('purchased', args=()))
-			elif type == 'display':
+			if type == 'purchased':
+			 	person.purchased.remove(article)
+			 	person.save()
+			 	return HttpResponseRedirect(reverse('purchased', args=()))
+			if type == 'display':
 				person.display.remove(article)
 				person.save()
 				article.delete()
 				return HttpResponseRedirect(reverse('display', args=()))
-			else:
-				return HttpResponseRedirect(reverse('search', args=('article', )))
+			
+			return HttpResponseRedirect(reverse('search', args=('article', )))
 		except Article.DoesNotExist:
 			return render(request, "circle/error.html", context={"message": "No Article Found.!!", "type": "Type Error", "link": "search"})
 		except Person.DoesNotExist:
